Extract shared fiducial helpers into `marker_pose.py`
